+++ R/rds.R
@@ -1,6 +1,0 @@
-rds <- function(x, compress = "xz") {
-  tf <- tempfile(fileext = ".RData")
-  save(x, file = tf, compress = compress, version = 2)
-
-  tf
-}