+++ src/init.c
@@ -1,8 +1,6 @@
 #include <Rinternals.h>
 #include <R_ext/Rdynload.h>
 
-
-
 /* .Call calls */
 extern SEXP C_addr(SEXP x);
 
@@ -11,7 +9,7 @@ static const R_CallMethodDef CallEntries[] = {
     {NULL, NULL, 0}
 };
 
-void R_init_addr(DllInfo *dll) {
+void R_init_dsn(DllInfo *dll) {
     R_registerRoutines(dll, NULL, CallEntries, NULL, NULL);
     R_useDynamicSymbols(dll, FALSE);
 }